==> 2.RandomOverSampling.ipynb <==
!pip install pyspark
!apt install openjdk-8-jdk-headless -qq

import pyspark
from pyspark.sql import *
from pyspark.sql.types import *
import pyspark.sql.functions as F
from pyspark import SparkContext, SparkConf


sc=pyspark.SparkContext()
# spark = SparkSession.builder.getOrCreate()

spark = SparkSession\
    .builder\
    .config("spark.driver.memory", "16g")\
    .config("spark.executor.memory", "4g")\
    .getOrCreate()


import matplotlib.pyplot as plt
import numpy as np
import random

from google.colab import drive
drive.mount('/content/drive')

train_data = spark.read.csv("/content/drive/MyDrive/빅데이터 처리 및 응용/팀플/train_data.csv", header=True, inferSchema=True)
test_data = spark.read.csv("/content/drive/MyDrive/빅데이터 처리 및 응용/팀플/test_data.csv", header=True, inferSchema=True)


# RandomOverSampling 알고리즘 구현

# 주 클래스와 소수 클래스의 비율 계산
origin_major_df = train_data.filter(train_data.HeartDisease_Index == 0.0)
origin_minor_df = train_data.filter(train_data.HeartDisease_Index == 1.0)

ratio = int(origin_major_df.count()/origin_minor_df.count())
print("ratio: {}".format(ratio))
a = range(ratio)

# 소수 클래스 복제
major_df = train_data.filter(train_data.HeartDisease_Index == 0.0)
minor_df = train_data.filter(train_data.HeartDisease_Index == 1.0)

create_data= minor_df.withColumn("dummy", F.explode(F.array([F.lit(x) for x in a]))).drop('dummy')

# 다수 클래스와 병합
oversample_train_data = major_df.unionAll(create_data)

oversample_train_data.show()

oversample_major_df = oversample_train_data.filter(oversample_train_data.HeartDisease_Index == 0.0)
oversample_minor_df = oversample_train_data.filter(oversample_train_data.HeartDisease_Index == 1.0)

# 데이터 비율 및 수
print("샘플링 전 데이터 개수: {}".format(train_data.count()))
print("샘플링 전 0인 경우 : {}".format(major_df.count()))
print("샘플링 전 1인 경우 : {}".format(minor_df.count()))
print()
print()
print("오버 샘플링 후 데이터 전체수: {}".format(oversample_train_data.count()))
print("오버 샘플링 데이터 label 0인 경우 : {}".format(oversample_major_df.count()))
print("오버 샘플링 데이터 label 1인 경우 : {}".format(oversample_minor_df.count()))

# RandomOverSampling으로 데이터 증강한 결과 내보내기

oversample_train_data.toPandas().to_csv('/content/drive/MyDrive/빅데이터 처리 및 응용/팀플/oversample_train_data.csv', index=False, encoding='CP949')
